refactor(main-api): event 패키지 이동
