+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/domain/device/active/ActiveDeviceFilter.java
@@ -7,7 +7,7 @@ import com.whoz_in.domain.network_log.MonitorLog;
 import com.whoz_in.domain.network_log.MonitorLogRepository;
 import com.whoz_in.main_api.query.device.application.active.view.ActiveDevice;
 import com.whoz_in.main_api.query.device.application.active.view.ActiveDeviceViewer;
-import com.whoz_in.main_api.query.device.application.active.event.ActiveDeviceFinded;
+import com.whoz_in.main_api.shared.domain.device.active.event.ActiveDeviceFinded;
 import com.whoz_in.main_api.shared.event.Events;
 import java.util.List;
 import java.util.Objects;
