+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/domain/device/active/InActiveDeviceFilter.java
@@ -9,7 +9,7 @@ import com.whoz_in.domain.network_log.MonitorLog;
 import com.whoz_in.domain.network_log.MonitorLogRepository;
 import com.whoz_in.main_api.query.device.application.active.view.ActiveDevice;
 import com.whoz_in.main_api.query.device.application.active.view.ActiveDeviceViewer;
-import com.whoz_in.main_api.query.device.application.active.event.InActiveDeviceFinded;
+import com.whoz_in.main_api.shared.domain.device.active.event.InActiveDeviceFinded;
 import com.whoz_in.main_api.shared.event.Events;
 import java.time.Duration;
 import java.time.LocalDateTime;
